Correcting names and texts
